Added stuff. Changed use cases

+++ src/Account/Account.java
@@ -1,25 +1,41 @@
 package Account;
 import User.Client;
 
+
 public class Account implements AccountInterface {
-    private int balance;
+    private float balance;
     private Client client;
     private String accountNumber;
 
+
+    /**
+     * Creates new account for client with account number
+     * @param client New user client
+     * @param accountNumber Account number assigned to this account
+     */
     public Account(Client client, String accountNumber){
         balance = 0;
         this.client = client;
         this.accountNumber = accountNumber;
     }
 
+    /**
+     * Deposits amount into the current account
+     * @param amount Integer
+     */
     @Override
-    public void deposit(int amount) {
-        balance+=amount;
+    public void deposit(float amount) {
+        balance+=amount; // Could change this from int to float/double
 
     }
 
+    /**
+     * Withdraws amount from current account balance if balance is greater than the amount requested.
+     * Otherwise does nothing.
+     * @param amount
+     */
     @Override
-    public void withdraw(int amount) {
+    public void withdraw(float amount) {
         if(isSufficientFunds(amount)){
             balance-=amount;
         }
@@ -27,16 +43,25 @@ public class Account implements AccountInterface {
 
     }
 
+    /**
+     * Returns boolean if the account has sufficient funds for the amount requested.
+     * @param amount
+     * @return boolean
+     */
     @Override
-    public boolean isSufficientFunds(int amount) {
+    public boolean isSufficientFunds(float amount) {
         if(balance-amount>0){
             return true;
         }
         else return false;
     }
 
+    /**
+     * Returns the balance of the account
+     * @return balance
+     */
     @Override
-    public int getBalance() {
+    public float getBalance() {
         return balance;
     }
 }
